Scenarios Cleanup
