+++ gym_art/quadrotor_multi/scenarios/dynamic_diff_goal.py
@@ -1,0 +1,50 @@
+import numpy as np
+
+from gym_art.quadrotor_multi.scenarios.utils import get_z_value
+from gym_art.quadrotor_multi.scenarios.base import QuadrotorScenario
+
+
+class Scenario_dynamic_diff_goal(QuadrotorScenario):
+    def __init__(self, quads_mode, envs, num_agents, room_dims, room_dims_callback, rew_coeff, quads_formation,
+                 quads_formation_size):
+        super().__init__(quads_mode, envs, num_agents, room_dims, room_dims_callback, rew_coeff, quads_formation,
+                         quads_formation_size)
+        # teleport every [4.0, 6.0] secs
+        duration_time = 5.0
+        self.control_step_for_sec = int(duration_time * self.envs[0].control_freq)
+
+    def update_goals(self):
+        # Reset formation and related parameters
+        self.update_formation_and_relate_param()
+
+        # Reset goals
+        self.goals = self.generate_goals(num_agents=self.num_agents, formation_center=self.formation_center,
+                                         layer_dist=self.layer_dist)
+        np.random.shuffle(self.goals)
+
+    def step(self, infos, rewards, pos):
+        tick = self.envs[0].tick
+        if tick % self.control_step_for_sec == 0 and tick > 0:
+            box_size = self.envs[0].box
+            x, y = np.random.uniform(low=-box_size, high=box_size, size=(2,))
+
+            # Get z value, and make sure all goals will above the ground
+            z = get_z_value(num_agents=self.num_agents, num_agents_per_layer=self.num_agents_per_layer,
+                            box_size=box_size, formation=self.formation, formation_size=self.formation_size)
+
+            self.formation_center = np.array([x, y, z])
+            self.update_goals()
+
+            # Update goals to envs
+            for i, env in enumerate(self.envs):
+                env.goal = self.goals[i]
+
+        return infos, rewards
+
+    def reset(self):
+        # Update duration time
+        duration_time = np.random.uniform(low=4.0, high=6.0)
+        self.control_step_for_sec = int(duration_time * self.envs[0].control_freq)
+
+        # Reset formation, and parameters related to the formation; formation center; goals
+        self.standard_reset()
