+++ gym_art/quadrotor_multi/scenarios/dynamic_same_goal.py
@@ -1,0 +1,39 @@
+import numpy as np
+
+from gym_art.quadrotor_multi.scenarios.base import QuadrotorScenario
+
+
+class Scenario_dynamic_same_goal(QuadrotorScenario):
+    def __init__(self, quads_mode, envs, num_agents, room_dims, room_dims_callback, rew_coeff, quads_formation,
+                 quads_formation_size):
+        super().__init__(quads_mode, envs, num_agents, room_dims, room_dims_callback, rew_coeff, quads_formation,
+                         quads_formation_size)
+        # teleport every [4.0, 6.0] secs
+        duration_time = 5.0
+        self.control_step_for_sec = int(duration_time * self.envs[0].control_freq)
+
+    def update_formation_size(self, new_formation_size):
+        pass
+
+    def step(self, infos, rewards, pos):
+        tick = self.envs[0].tick
+        if tick % self.control_step_for_sec == 0 and tick > 0:
+            box_size = self.envs[0].box
+            x, y = np.random.uniform(low=-box_size, high=box_size, size=(2,))
+            z = np.random.uniform(low=-0.5 * box_size, high=0.5 * box_size) + 2.0
+            z = max(0.25, z)
+            self.formation_center = np.array([x, y, z])
+            self.goals = self.generate_goals(num_agents=self.num_agents, formation_center=self.formation_center,
+                                             layer_dist=0.0)
+            for i, env in enumerate(self.envs):
+                env.goal = self.goals[i]
+
+        return infos, rewards
+
+    def reset(self):
+        # Update duration time
+        duration_time = np.random.uniform(low=4.0, high=6.0)
+        self.control_step_for_sec = int(duration_time * self.envs[0].control_freq)
+
+        # Reset formation, and parameters related to the formation; formation center; goals
+        self.standard_reset()
