+++ gym_art/quadrotor_multi/scenarios/ep_rand_bezier.py
@@ -1,0 +1,51 @@
+import numpy as np
+import bezier
+
+from gym_art.quadrotor_multi.scenarios.base import QuadrotorScenario
+
+
+class Scenario_ep_rand_bezier(QuadrotorScenario):
+    def step(self, infos, rewards, pos):
+        # randomly sample new goal pos in free space and have the goal move there following a bezier curve
+        tick = self.envs[0].tick
+        control_freq = self.envs[0].control_freq
+        num_secs = 5
+        control_steps = int(num_secs * control_freq)
+        t = tick % control_steps
+        room_dims = np.array(self.room_dims) - self.formation_size
+        # min and max distance the goal can spawn away from its current location. 30 = empirical upper bound on
+        # velocity that the drones can handle.
+        max_dist = min(30, max(room_dims))
+        min_dist = max_dist / 2
+        if tick % control_steps == 0 or tick == 1:
+            # sample a new goal pos that's within the room boundaries and satisfies the distance constraint
+            new_goal_found = False
+            while not new_goal_found:
+                low, high = np.array([-room_dims[0] / 2, -room_dims[1] / 2, 0]), np.array(
+                    [room_dims[0] / 2, room_dims[1] / 2, room_dims[2]])
+                # need an intermediate point for a deg=2 curve
+                new_pos = np.random.uniform(low=-high, high=high, size=(2, 3)).reshape(3, 2)
+                # add some velocity randomization = random magnitude * unit direction
+                new_pos = new_pos * np.random.randint(min_dist, max_dist + 1) / np.linalg.norm(new_pos, axis=0)
+                new_pos = self.goals[0].reshape(3, 1) + new_pos
+                lower_bound = np.expand_dims(low, axis=1)
+                upper_bound = np.expand_dims(high, axis=1)
+                new_goal_found = (new_pos > lower_bound + 0.5).all() and (
+                        new_pos < upper_bound - 0.5).all()  # check bounds that are slightly smaller than the room dims
+            nodes = np.concatenate((self.goals[0].reshape(3, 1), new_pos), axis=1)
+            nodes = np.asfortranarray(nodes)
+            pts = np.linspace(0, 1, control_steps)
+            curve = bezier.Curve(nodes, degree=2)
+            self.interp = curve.evaluate_multi(pts)
+            # self.interp = np.clip(self.interp, a_min=np.array([0,0,0.2]).reshape(3,1), a_max=high.reshape(3,
+            # 1)) # want goal clipping to be slightly above the floor
+        if tick % control_steps != 0 and tick > 1:
+            self.goals = np.array([self.interp[:, t] for _ in range(self.num_agents)])
+
+            for i, env in enumerate(self.envs):
+                env.goal = self.goals[i]
+
+        return infos, rewards
+
+    def update_formation_size(self, new_formation_size):
+        pass
