+++ gym_art/quadrotor_multi/scenarios/static_diff_goal.py
@@ -1,0 +1,6 @@
+from gym_art.quadrotor_multi.scenarios.base import QuadrotorScenario
+
+
+class Scenario_static_diff_goal(QuadrotorScenario):
+    def step(self, infos, rewards, pos):
+        return infos, rewards
