+++ gym_art/quadrotor_multi/scenarios/static_same_goal.py
@@ -1,0 +1,9 @@
+from gym_art.quadrotor_multi.scenarios.base import QuadrotorScenario
+
+
+class Scenario_static_same_goal(QuadrotorScenario):
+    def update_formation_size(self, new_formation_size):
+        pass
+
+    def step(self, infos, rewards, pos):
+        return infos, rewards
