+++ gym_art/quadrotor_multi/scenarios/swap_goals.py
@@ -1,0 +1,34 @@
+import numpy as np
+
+from gym_art.quadrotor_multi.scenarios.base import QuadrotorScenario
+
+
+class Scenario_swap_goals(QuadrotorScenario):
+    def __init__(self, quads_mode, envs, num_agents, room_dims, room_dims_callback, rew_coeff, quads_formation,
+                 quads_formation_size):
+        super().__init__(quads_mode, envs, num_agents, room_dims, room_dims_callback, rew_coeff, quads_formation,
+                         quads_formation_size)
+        # teleport every [4.0, 6.0] secs
+        duration_time = 5.0
+        self.control_step_for_sec = int(duration_time * self.envs[0].control_freq)
+
+    def update_goals(self):
+        np.random.shuffle(self.goals)
+        for env, goal in zip(self.envs, self.goals):
+            env.goal = goal
+
+    def step(self, infos, rewards, pos):
+        tick = self.envs[0].tick
+        # Switch every [4, 6] seconds
+        if tick % self.control_step_for_sec == 0 and tick > 0:
+            self.update_goals()
+
+        return infos, rewards
+
+    def reset(self):
+        # Update duration time
+        duration_time = np.random.uniform(low=4.0, high=6.0)
+        self.control_step_for_sec = int(duration_time * self.envs[0].control_freq)
+
+        # Reset formation, and parameters related to the formation; formation center; goals
+        self.standard_reset()
